refactor(group): move group-related classes into dedicated package
