+++ tracker-api/src/main/java/com/sloyardms/trackerapi/group/GroupService.java
@@ -1,8 +1,8 @@
-package com.sloyardms.trackerapi.service.interfaces;
+package com.sloyardms.trackerapi.group;
 
-import com.sloyardms.trackerapi.dto.GroupCreateDto;
-import com.sloyardms.trackerapi.dto.GroupDto;
-import com.sloyardms.trackerapi.dto.GroupUpdateDto;
+import com.sloyardms.trackerapi.group.dto.GroupCreateDto;
+import com.sloyardms.trackerapi.group.dto.GroupDto;
+import com.sloyardms.trackerapi.group.dto.GroupUpdateDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
