+++ tracker-api/src/main/java/com/sloyardms/trackerapi/group/GroupServiceImpl.java
@@ -1,16 +1,14 @@
-package com.sloyardms.trackerapi.service;
+package com.sloyardms.trackerapi.group;
 
-import com.sloyardms.trackerapi.dto.GroupCreateDto;
-import com.sloyardms.trackerapi.dto.GroupDto;
-import com.sloyardms.trackerapi.dto.GroupUpdateDto;
-import com.sloyardms.trackerapi.entity.Group;
+import com.sloyardms.trackerapi.group.dto.GroupCreateDto;
+import com.sloyardms.trackerapi.group.dto.GroupDto;
+import com.sloyardms.trackerapi.group.dto.GroupUpdateDto;
+import com.sloyardms.trackerapi.group.entity.Group;
 import com.sloyardms.trackerapi.user.entity.User;
 import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
 import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
-import com.sloyardms.trackerapi.mapper.GroupMapper;
-import com.sloyardms.trackerapi.repository.GroupRepository;
-import com.sloyardms.trackerapi.repository.UserRepository;
-import com.sloyardms.trackerapi.service.interfaces.GroupService;
+import com.sloyardms.trackerapi.group.mapper.GroupMapper;
+import com.sloyardms.trackerapi.user.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
