+++ tracker-api/src/main/java/com/sloyardms/trackerapi/group/dto/GroupCreateDto.java
@@ -1,21 +1,20 @@
-package com.sloyardms.trackerapi.dto;
+package com.sloyardms.trackerapi.group.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.UUID;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class GroupDto {
+public class GroupCreateDto {
 
-    private UUID uuid;
-    private UUID userUuid;
+    @NotBlank
     private String name;
+
     private String description;
 
 }
