+++ tracker-api/src/main/java/com/sloyardms/trackerapi/group/entity/Group.java
@@ -1,5 +1,6 @@
-package com.sloyardms.trackerapi.entity;
+package com.sloyardms.trackerapi.group.entity;
 
+import com.sloyardms.trackerapi.entity.Auditable;
 import com.sloyardms.trackerapi.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -17,7 +18,7 @@ import java.util.UUID;
 @Table(name = "groups", uniqueConstraints = {
         @UniqueConstraint(columnNames = {"user_uuid", "name"})
 })
-public class Group extends Auditable{
+public class Group extends Auditable {
 
     @Id
     @Column(name = "group_uuid", nullable = false, updatable = false)
